% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo-autoplot.R
\name{autoplot.loo}
\alias{autoplot.loo}
\title{Diagnostic Plot(s) for PSIS in ggplot2}
\usage{
\method{autoplot}{loo}(object, diagnostic = c("k", "n_eff"), ...)
}
\arguments{
\item{object}{object created by \code{\link[loo:loo]{loo::loo()}} or \code{\link[loo:psis]{loo::psis()}}.}

\item{diagnostic}{choice. Should the Pareto K estimates or PSIS effective
sample size be plotted.}

\item{...}{Currently unused, may be forwarded to a \code{geom} in the future.}
}
\description{
A ggplot2 recreation of the \code{\link[loo:pareto-k-diagnostic]{plot()}} methods
for \code{loo} package objects.
}
