% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{compare_via_loglik}
\alias{compare_via_loglik}
\title{Compare Models using Log Likelihood}
\usage{
compare_via_loglik(loglik_list, cols, type = c("raw", "gt"))
}
\arguments{
\item{loglik_list}{named list. Each element is log likelihood of the model,
stored in a \code{\link[mcmcr:mcmcarray-object]{mcmcarray}}.}

\item{cols}{character. Columns to be used in \code{\link[loo:loo_compare]{loo::loo_compare()}}.}

\item{type}{character choice. Defines output type.}
}
\description{
High-level wrapper around chosen comparison metrics, calculated per chain.
}
