% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete-jags-fit.R
\name{complete_jags_fit}
\alias{complete_jags_fit}
\title{Complete JAGS Fit}
\usage{
complete_jags_fit(file, data, inits, jags_n, varnames)
}
\arguments{
\item{file, data, inits}{Passed to \code{\link[rjags:jags.model]{rjags::jags.model()}}.}

\item{jags_n}{named list. TODO.}

\item{varnames}{character. Variable names to be monitored via JAGS.}
}
\value{
list with
\itemize{
\item \code{data}: same as input.
\item \code{model}: object returned from \code{\link[rjags:jags.model]{jags.model}}.
\item \code{output}: samples stored in a \code{\link[mcmcr:mcmcarray-object]{mcmcarray}}.
\item \code{jags_n}: same as input.
}
}
\description{
Fits a model in JAGS including compilation, warm-up and sampling.
}
