% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws-constructors.R
\name{draws}
\alias{draws}
\alias{draws.mcmcr}
\alias{draws.mcmcarray}
\title{Extract Draws into a Tidy Data Format}
\usage{
draws(x, ...)

\method{draws}{mcmcr}(x, pars = NULL, ...)

\method{draws}{mcmcarray}(x, name = "par", ...)
}
\arguments{
\item{x}{object used to select a method.}

\item{...}{extra arguments ignored.}

\item{pars}{character vector. Names of parameters to subset by.}

\item{name}{character. Parameter name to be assumed for the \code{.term} column.}
}
\description{
Mimics the behaviour of \code{tidybayes} using a \code{.term} column
}
