% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws-methods.R
\name{draws_add_truth}
\alias{draws_add_truth}
\title{Draws, Add a 'Truth' Column}
\usage{
draws_add_truth(draws, ..., .list = NULL)
}
\arguments{
\item{draws}{See \strong{Draws} section below.}

\item{...}{Name-value pairs to be joined to the table.}

\item{.list}{list. Alternative way of passing arguments to \code{...}.}
}
\description{
Adds column based on a supplied R object for comparison.
}
\section{Draws}{

Typically output from \code{\link[=gather_draws_by_term]{gather_draws_by_term()}}, a \code{draws}-style
\link[tibble:tibble]{tibble} assumed to have columns:
\itemize{
\item \code{.chain}, \code{.iteration}, \code{.draw}. MCMC sample indices.
\item \code{.term}. Parameter information.
\item \code{.value}. MCMC sample value.
}
}

\seealso{
Other draws methods: 
\code{\link{draws-to-drawsvis}},
\code{\link{draws_separate_chains}()},
\code{\link{draws_summarise}()},
\code{\link{draws_unnest_term}()},
\code{\link{ggdraws}}
}
\concept{draws methods}
