% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{draws_example}
\alias{draws_example}
\title{Example Draws Object(s)}
\format{
List of three tibbles with columns
\itemize{
\item \code{.chain}, \code{.iteration}, \code{.draw}. MCMC sample indices.
\item \code{.term}. Parameter information.
\item \code{.value}. MCMC sample value.
}
}
\usage{
draws_example
}
\description{
Draws-type tibbles for the \link[=mcmcr_example]{mcmcr::mcmcr_example}.
}
\keyword{datasets}
