% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws-methods.R
\name{draws_separate_chains}
\alias{draws_separate_chains}
\title{Draws, Separate Chains}
\usage{
draws_separate_chains(draws, .keep = FALSE)
}
\arguments{
\item{draws}{object of class \code{draws}.}

\item{.keep}{logical. When \code{TRUE}, the chain column is kept in the output.}
}
\description{
Split Draws-type tibble into a list of \code{draws} for chain-by-chain
manipulation.
}
\section{Draws}{

Typically output from \code{\link[=draws]{draws()}}, extends \link[tibble:tibble]{tibble} class
and is assumed to have columns:
\itemize{
\item \code{.chain}, \code{.iteration}, \code{.draw} (MCMC sample indices),
\item \code{.term} (parameter information),
\item \code{.value} (MCMC sample value).
}
}

\seealso{
Other draws methods: 
\code{\link{draws_add_truth}()},
\code{\link{draws_summarise}()},
\code{\link{draws_unnest_term}()},
\code{\link{ggdraws}}
}
\concept{draws methods}
