% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws-methods.R
\name{draws_summarise}
\alias{draws_summarise}
\title{Draws, Summarise}
\usage{
draws_summarise(draws, .point = mean, .width = 0.95, collapse_chains = FALSE)
}
\arguments{
\item{draws}{object of class \code{draws}.}

\item{.point}{function. Point estimate method, defaults to \code{\link{mean}}.}

\item{.width}{numeric. Width of CI to be calculated with \code{\link[stats:quantile]{stats::quantile}}.}

\item{collapse_chains}{logical. Single estimate for all chains when \code{TRUE}.}
}
\description{
Calculate point and interval summaries of draws.
}
\section{Draws}{

Typically output from \code{\link[=draws]{draws()}}, extends \link[tibble:tibble]{tibble} class
and is assumed to have columns:
\itemize{
\item \code{.chain}, \code{.iteration}, \code{.draw} (MCMC sample indices),
\item \code{.term} (parameter information),
\item \code{.value} (MCMC sample value).
}
}

\seealso{
\code{\link[ggdist:point_interval]{ggdist::point_interval()}} for inspiration.

Other draws methods: 
\code{\link{draws_add_truth}()},
\code{\link{draws_separate_chains}()},
\code{\link{draws_unnest_term}()},
\code{\link{ggdraws}}
}
\concept{draws methods}
