% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws-unnest-term.R
\name{draws_unnest_term}
\alias{draws_unnest_term}
\title{Draws, Unnest Term}
\usage{
draws_unnest_term(draws, .names)
}
\arguments{
\item{draws}{tibble. See \strong{Draws} section below.}

\item{.names}{character. Dimension colnames, defaults to \verb{dim_*}.}
}
\description{
Splits the \code{.term} column of a draws object into a data frame column.
}
\section{Draws}{

Typically output from \code{\link[=gather_draws_by_term]{gather_draws_by_term()}}, a \code{draws}-style
\link[tibble:tibble]{tibble} assumed to have columns:
\itemize{
\item \code{.chain}, \code{.iteration}, \code{.draw}. MCMC sample indices.
\item \code{.term}. Parameter information.
\item \code{.value}. MCMC sample value.
}
}

\seealso{
Other draws methods: 
\code{\link{draws_separate_chains}()}
}
\concept{draws methods}
