% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawsvis-methods.R
\name{drawsvis_add_term_subtitle}
\alias{drawsvis_add_term_subtitle}
\title{Drawsvis, Add Subtitle(s)}
\usage{
drawsvis_add_term_subtitle(dv, ...)
}
\arguments{
\item{dv}{Drawsvis-type tibble. See \strong{Drawsvis} section.}

\item{...}{Modifiers where position implies dimension.
\itemize{
\item Character vectors are used as extractor functions on each index.
\item Functions or formulas are applied to each index.
\item Explicit missing arguments mean that dimension is omitted from the label.
\item \code{NULL} values assert the default and indices are used as integers.
}}
}
\description{
Drawsvis, Add Subtitle(s)
}
\section{Drawsvis}{

Nested version of \code{\link{draws}} that also extends \link[tibble:tibble]{tibble}.
Has one plot per row instead of one draw per row, columns are typically:
\itemize{
\item \code{.chain} (MCMC chain index),
\item \code{.term} (parameter information),
\item \code{.plot} (list column of MCMC visualisations).
}
}

\seealso{
Other drawsvis methods: 
\code{\link{drawsvis_to_patchwork}()}
}
\concept{drawsvis methods}
