% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drawsvis_example}
\alias{drawsvis_example}
\title{Example Drawsvis Object(s)}
\format{
List of three tibbles with columns
\itemize{
\item \code{.chain}. MCMC chain index.
\item \code{.term}. Parameter information.
\item \code{.plot}. MCMC visualisation.
}
}
\usage{
drawsvis_example
}
\description{
Drawsvis-type tibbles for the \code{\link[mcmcr:mcmcr_example]{mcmcr_example}}.
}
\keyword{datasets}
