% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enterm.R
\name{enterm}
\alias{enterm}
\title{Convert Vector to Data Frame with Term(s)}
\usage{
enterm(x, name = "x", .value = ".value")
}
\arguments{
\item{x}{object. Typically a scalar, vector or matrix.}

\item{name}{character. Name to be assumed in the \code{.term} column.}

\item{.value}{character. Name of the column containing the values.}
}
\description{
Convert Vector to Data Frame with Term(s)
}
