% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates-per-chain.R
\name{estimates_per_chain}
\alias{estimates_per_chain}
\title{Estimates per Chain}
\usage{
estimates_per_chain(x, fun, as_df, ...)
}
\arguments{
\item{x}{\code{\link[mcmcr:mcmcarray-object]{mcmcarray}} object.}

\item{fun}{A function that given a numeric vector returns a numeric scalar.}

\item{as_df}{A flag indicating whether to return the results as a
data frame versus a named list.}

\item{...}{Optional arguments to fun.}
}
\description{
Calculate the estimates of an MCMC object on a chain-per-chain basis.
}
\seealso{
Other mcmcarray methods: 
\code{\link{gather_draws_by_term}()}
}
\concept{mcmcarray methods}
