% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather-draws-by-term.R
\name{gather_draws_by_term}
\alias{gather_draws_by_term}
\alias{gather_draws_by_term.mcmcr}
\alias{gather_draws_by_term.mcmcarray}
\title{Extract Draws into a Tidy Data Format}
\usage{
gather_draws_by_term(x, ...)

\method{gather_draws_by_term}{mcmcr}(x, ...)

\method{gather_draws_by_term}{mcmcarray}(x, name = "par", ...)
}
\arguments{
\item{x}{\code{\link[mcmcr:mcmcarray-object]{mcmcarray}} object.}

\item{...}{extra arguments ignored.}

\item{name}{character. Parameter name to be assumed for the \code{.term} column.}
}
\description{
Mimics the behaviour of \code{\link[tidybayes:spread_draws]{tidybayes::gather_draws()}} using a \code{.term} column
}
\seealso{
Other mcmcarray methods: 
\code{\link{estimates_per_chain}()}
}
\concept{mcmcarray methods}
