% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdraws.R
\name{ggdraws}
\alias{ggdraws}
\alias{ggdraws_trace}
\alias{ggdraws_dens}
\alias{ggdraws_acf}
\alias{ggdraws_raster}
\title{Draws Visualisations}
\usage{
ggdraws_trace(draws, mapping = NULL, ...)

ggdraws_dens(draws, mapping = NULL, ...)

ggdraws_acf(draws, mapping = NULL, ..., lag.max)

ggdraws_raster(draws, mapping = NULL, ..., levels)
}
\arguments{
\item{draws}{See \strong{Draws} section below.}

\item{mapping}{Optional \code{aes}. Updates the mapping in the relevant \code{geom}.}

\item{...}{extra arguments passed to relevant \code{geom}.}

\item{lag.max}{integer. Passed to \code{\link[stats:acf]{acf}}.}

\item{levels}{vector. Range of possible values needed when \code{.value} is
converted to a factor.}
}
\description{
Various methods to convert a Draws-type tibble into a ggplot2 visualisation.
}
\section{Draws}{

Typically output from \code{\link[=gather_draws_by_term]{gather_draws_by_term()}}, a \code{draws}-style
\link[tibble:tibble]{tibble} assumed to have columns:
\itemize{
\item \code{.chain}, \code{.iteration}, \code{.draw}. MCMC sample indices.
\item \code{.term}. Parameter information.
\item \code{.value}. MCMC sample value.
}
}

\seealso{
Other draws methods: 
\code{\link{draws-to-drawsvis}},
\code{\link{draws_add_truth}()},
\code{\link{draws_separate_chains}()},
\code{\link{draws_summarise}()},
\code{\link{draws_unnest_term}()}
}
\concept{draws methods}
