% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jags-n-data}
\alias{jags-n-data}
\alias{jags_n_debug}
\alias{jags_n_short}
\alias{jags_n_thin}
\title{Example JAGS Algorithm Controls}
\format{
A \code{list} with elements: \code{n.adapt}, \code{n.update}, \code{n.iter}, \code{n.thin}
and \code{n.chains}.

An object of class \code{jags_n} of length 5.

An object of class \code{jags_n} of length 5.

An object of class \code{jags_n} of length 5.
}
\usage{
jags_n_debug

jags_n_short

jags_n_thin
}
\description{
Various controls to be used in JAGS sampling
}
\section{Functions}{
\itemize{
\item \code{jags_n_debug}: Used for code testing and verification.

\item \code{jags_n_short}: Used for analysis stage on full data, fewer samples.

\item \code{jags_n_thin}: Used for analysis stage \strong{with} thinning.
}}

\keyword{datasets}
