% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete-jags-fit.R
\name{jags_inits}
\alias{jags_inits}
\title{JAGS Initial Values}
\usage{
jags_inits(seed_base, .RNG.name = "base::Wichmann-Hill", ...)
}
\arguments{
\item{seed_base}{numeric. Seed value to base the entire run on.}

\item{.RNG.name}{character. JAGS RNG method.}

\item{...}{Used for initial parameter values also given in \code{inits}.}
}
\description{
Generate a function for \code{inits} in \link[rjags:jags.model]{rjags::jags.model} with a chain-varying,
reproducible RNG scheme. Each seed equals \code{seed_base} times chain index.
}
