% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags-n.R
\name{jags_n}
\alias{jags_n}
\title{Create JAGS Algorithm Controls}
\usage{
jags_n(n.adapt, n.update, n.iter, n.thin, n.chains)
}
\arguments{
\item{n.adapt}{integer. Number of samples discarded in adaptation stage;}

\item{n.update}{integer. Number of samples discarded in warm-up stage;}

\item{n.iter}{integer. Number of samples stored from the monitoring stage;}

\item{n.thin}{integer. Thinning interval, only every \emph{n.thin}-th sample kept;}

\item{n.chains}{integer. Number of chains to create.}
}
\value{
Object with S3 class \code{jags_n} that can be used as a normal \code{list}.
}
\description{
Named list to control how long JAGS samples from the posterior.
}
