% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\alias{log_likelihood}
\title{Likelihood Generic (MCMC)}
\usage{
likelihood(object, ...)

log_likelihood(object, ...)
}
\arguments{
\item{object}{object used to select a method.}

\item{...}{extra arguments passed to method.}
}
\value{
An \code{\link[mcmcr:mcmcarray-object]{mcmcarray}} of a \code{n_obs}-length vector.
}
\description{
Get likelihood contribution of each observation over each MCMC sample.
}
