% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo-mixexpert}
\alias{loo-mixexpert}
\alias{log_pred_density}
\alias{waic.mcmcarray}
\alias{loo.mcmcarray}
\title{LOO Cross Validation within \code{mcmcr}}
\usage{
log_pred_density(x)

waic.mcmcarray(x, ...)

loo.mcmcarray(x, ...)
}
\arguments{
\item{x}{\code{\link[mcmcr:mcmcarray-object]{mcmcarray}} object. Log-likelihood.}

\item{...}{extra arguments passed to and from other methods.}
}
\description{
Methods for \code{\link[loo:loo-package]{loo()}} package functionality.
}
\section{Functions}{
\itemize{
\item \code{log_pred_density()}: Computes simulation-estimate for log pointwise predictive density (LPD).

}}
\references{
Vehtari et al. (2017), \url{https://doi.org/10.1007/s11222-016-9696-4}.
}
\seealso{
LPD (see references), PSIS (\code{\link[loo:loo]{loo::loo()}}),
WAIC (\code{\link[loo:waic]{loo::waic()}}).
}
