% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-mcmcarray.R
\name{map_mcmcarray}
\alias{map_mcmcarray}
\alias{map2_mcmcarray}
\alias{pmap_mcmcarray}
\title{Apply function to each MCMC Draw}
\usage{
map_mcmcarray(x, .f, ...)

map2_mcmcarray(x, y, .f, ...)

pmap_mcmcarray(l, .f, ...)
}
\arguments{
\item{x, y, l}{\href{mcmcr::mcmcarray-object()}{mcmcarray} object(s), \code{l} is a
list of such object(s).}

\item{.f}{function or formula. See \code{\link[rlang:as_function]{rlang::as_function()}}.}

\item{...}{extra arguments passed to \code{.f}}
}
\value{
\href{mcmcr::mcmcarray-object()}{mcmcarray} object, identical iteration
and chain dimensions, storing output of \code{.f(l, ...)} or similar.
}
\description{
Recreation of purrr's \code{map}-like functionality for \code{mcmcarray}(s).
}
