% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcarray-to-pred.R
\name{mcmcarray_to_pred}
\alias{mcmcarray_to_pred}
\title{MCMC Array to Prediction Tibble}
\usage{
mcmcarray_to_pred(x, collapse_chains = FALSE)
}
\arguments{
\item{x}{\code{\link[mcmcr:mcmcarray-object]{mcmcarray}}. Representing predictions.}

\item{collapse_chains}{logical. Single estimate for all chains when \code{TRUE}.}
}
\description{
Convert \code{n}-length MCMC array into nested predictions where each tibble in
the nest contains \code{.chain} and \code{.pred}, \code{.pred_lower}, \code{.pred_upper}.
}
