% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdraws.R
\name{modify_aes}
\alias{modify_aes}
\title{Modify \code{aes}}
\usage{
modify_aes(..., user)
}
\arguments{
\item{...}{developer supplied defaults.}

\item{user}{list. User supplied mappings.}
}
\description{
Set mapping defaults and update with user input, if supplied
}
\keyword{internal}
