% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdraws-parse.R
\name{parse_ggdraws}
\alias{parse_ggdraws}
\alias{parse_ggdraws.default}
\alias{parse_ggdraws.character}
\title{Parse ggdraws}
\usage{
parse_ggdraws(.ggdraws, ...)

\method{parse_ggdraws}{default}(.ggdraws, ...)

\method{parse_ggdraws}{character}(.ggdraws, ...)
}
\arguments{
\item{.ggdraws}{\strong{TODO}.}

\item{...}{\strong{TODO}.}
}
\description{
Parse user input into a chosen (or custom) visualisation function.
}
\keyword{internal}
