% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags-n.R
\name{print.jags_n}
\alias{print.jags_n}
\title{Show JAGS Controls}
\usage{
\method{print}{jags_n}(x, ...)
}
\arguments{
\item{x}{object of \code{\link{jags_n}} type.}

\item{...}{extra arguments silently ignored.}
}
\description{
Show JAGS Controls
}
