% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_x_iterations}
\alias{scale_x_iterations}
\title{Position Scale for MCMC Iterations}
\usage{
scale_x_iterations(..., n = 5)
}
\arguments{
\item{...}{passed to \code{scale_x_continuous}}

\item{n}{integer. Number of desirable breaks.}
}
\description{
Wrapper around \code{\link[ggplot2:scale_continuous]{scale_x_continuous}} with
updated defaults for thousands of MCMC iterations.
}
