% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws-separate-chains.R
\name{draws_separate_chains}
\alias{draws_separate_chains}
\title{Draws, Separate Chains}
\usage{
draws_separate_chains(draws, .keep = FALSE)
}
\arguments{
\item{draws}{See \strong{Draws} section below.}

\item{.keep}{logical. When \code{TRUE}, the chain column is kept in the output.}
}
\description{
Split Draws-type tibble into a list of \code{draws} for chain-by-chain
manipulation.
}
\section{Draws}{

Typically output from \code{\link[=gather_draws_by_term]{gather_draws_by_term()}}, a \code{draws}-style
\link[tibble:tibble]{tibble} assumed to have columns:
\itemize{
\item \code{.chain}, \code{.iteration}, \code{.draw}. MCMC sample indices.
\item \code{.term}. Parameter information.
\item \code{.value}. MCMC sample value.
}
}

\seealso{
Other draws methods: 
\code{\link{draws_add_truth}()},
\code{\link{draws_summarise}()},
\code{\link{draws_unnest_term}()},
\code{\link{ggdraws}}
}
\concept{draws methods}
