% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws-summarise.R
\name{draws_summarise}
\alias{draws_summarise}
\title{Draws, Summarise}
\usage{
draws_summarise(draws, ..., .point = mean)
}
\arguments{
\item{draws}{See \strong{Draws} section below.}

\item{...}{passed to \code{\link[ggdist:point_interval]{point_interval()}}.}

\item{.point}{function. Point estimate method (default changed to mean).}
}
\description{
Wrapper around \code{\link[ggdist:point_interval]{point_interval}} to calculate
point and interval summaries of draws.
}
\section{Draws}{

Typically output from \code{\link[=gather_draws_by_term]{gather_draws_by_term()}}, a \code{draws}-style
\link[tibble:tibble]{tibble} assumed to have columns:
\itemize{
\item \code{.chain}, \code{.iteration}, \code{.draw}. MCMC sample indices.
\item \code{.term}. Parameter information.
\item \code{.value}. MCMC sample value.
}
}

\seealso{
Other draws methods: 
\code{\link{draws-to-drawsvis}},
\code{\link{draws_add_truth}()},
\code{\link{draws_separate_chains}()},
\code{\link{draws_unnest_term}()},
\code{\link{ggdraws}}
}
\concept{draws methods}
