% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draws-methods.R
\name{draws_unnest_term}
\alias{draws_unnest_term}
\title{Draws, Unnest Term}
\usage{
draws_unnest_term(draws)
}
\arguments{
\item{draws}{object of class \code{draws}.}
}
\description{
Splits the \code{.term} column of a draws object into a data frame column.
}
\section{Draws}{

Typically output from \code{\link[=draws]{draws()}}, extends \link[tibble:tibble]{tibble} class
and is assumed to have columns:
\itemize{
\item \code{.chain}, \code{.iteration}, \code{.draw} (MCMC sample indices),
\item \code{.term} (parameter information),
\item \code{.value} (MCMC sample value).
}
}

\seealso{
Other draws methods: 
\code{\link{draws_add_truth}()},
\code{\link{draws_separate_chains}()},
\code{\link{draws_summarise}()},
\code{\link{ggdraws}}
}
\concept{draws methods}
