% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawsvis.R
\name{drawsvis}
\alias{drawsvis}
\title{Create Plotting Table (\code{drawsvis})}
\usage{
drawsvis(x, .ggdraws, ...)
}
\arguments{
\item{x}{object used to select a method.}

\item{.ggdraws}{see \strong{ggdraws} section}

\item{...}{extra arguments passed to invoked \code{.ggdraws} function.}
}
\description{
Generic used to convert an object to a \code{drawsvis} tibble.
}
\section{ggdraws}{

The role of the \code{.ggdraws} argument is invoke a function capable of
converting a (typically filtered) Draws-type tibble into a \code{ggplot2}
object.

For advanced users it can be a custom function (or formula) that should
take the signature \code{.f(draws, ...)}.

Alternatively, one can supply a character argument to invoke a built-in
method.
\itemize{
\item \code{"trace" -> ggdraws_trace}.
\item \code{"density" -> ggdraws_dens}.
\item \code{"acf" -> ggdraws_acf}.
\item \code{"raster" -> ggdraws_raster}.
}
}

