% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawsvis.R
\name{drawsvis}
\alias{drawsvis}
\title{Create \code{DrawsVis} tibble}
\usage{
drawsvis(draws, .ggdraws, ...)
}
\arguments{
\item{draws}{Draws-type tibble. See \strong{Draws} section below.}

\item{.ggdraws}{see \strong{ggdraws} section}

\item{...}{extra arguments passed to invoked \code{.ggdraws} function.}
}
\description{
Nests and already grouped Draws-type tibble into a Drawsvis-type tibble
with a list column of \code{ggplot2} objects.
}
\section{Draws}{

Typically output from \code{\link[=gather_draws_by_term]{gather_draws_by_term()}}, a \code{draws}-style
\link[tibble:tibble]{tibble} assumed to have columns:
\itemize{
\item \code{.chain}, \code{.iteration}, \code{.draw}. MCMC sample indices.
\item \code{.term}. Parameter information.
\item \code{.value}. MCMC sample value.
}
}

\section{ggdraws}{

The role of the \code{.ggdraws} argument is invoke a function capable of
converting a (typically filtered) Draws-type tibble into a \code{ggplot2}
object.

For advanced users it can be a custom function (or formula) that should
take the signature \code{.f(draws, ...)}.

Alternatively, one can supply a character argument to invoke a built-in
method.
\itemize{
\item \code{"trace" -> ggdraws_trace}.
\item \code{"density" -> ggdraws_dens}.
\item \code{"acf" -> ggdraws_acf}.
\item \code{"heatmap" -> ggdraws_heatmap}.
}
}

\seealso{
Other draws methods: 
\code{\link{draws_add_truth}()},
\code{\link{draws_separate_chains}()},
\code{\link{draws_summarise}()},
\code{\link{draws_unnest_term}()},
\code{\link{ggdraws}}
}
\concept{draws methods}
