% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawsvis.R
\name{drawsvis}
\alias{drawsvis}
\title{Create \code{drawsvis} tibble}
\usage{
drawsvis(draws, .ggdraws, ...)
}
\arguments{
\item{draws}{tibble. See \strong{Draws} section below.}

\item{.ggdraws}{\strong{TODO}.}

\item{...}{extra arguments passed to \code{.ggdraws}}
}
\description{
Group nests \code{draws} tibble into a \code{draws_vis} tibble.
}
\section{Draws}{

Typically output from \code{\link[=gather_draws_by_term]{gather_draws_by_term()}}, a \code{draws}-style
\link[tibble:tibble]{tibble} assumed to have columns:
\itemize{
\item \code{.chain}, \code{.iteration}, \code{.draw}. MCMC sample indices.
\item \code{.term}. Parameter information.
\item \code{.value}. MCMC sample value.
}
}

\seealso{
Other draws methods: 
\code{\link{draws_add_truth}()},
\code{\link{draws_separate_chains}()},
\code{\link{draws_summarise}()},
\code{\link{draws_unnest_term}()},
\code{\link{ggdraws}}
}
\concept{draws methods}
