% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawsvis-add-term-subtitle.R
\name{drawsvis_add_term_subtitle}
\alias{drawsvis_add_term_subtitle}
\title{Drawsvis Methods}
\usage{
drawsvis_add_term_subtitle(dv, ...)
}
\arguments{
\item{dv}{Drawsvis-type tibble. See \strong{Drawsvis} section below.}

\item{...}{Modifiers where position implies dimension.
\itemize{
\item Character vectors are used as extractor functions on each index.
\item Functions or formulas are applied to each index.
\item Explicit missing arguments mean that dimension is omitted from the label.
\item \code{NULL} values assert the default and indices are used as integers.
}}
}
\description{
Drawsvis Methods
}
\section{Drawsvis}{

Typically output from \code{\link[=drawsvis]{drawsvis()}}, a \code{drawsvis}-style
\link[tibble:tibble]{tibble} has one term per row instead of one draws
per row and is therefore assumed to have columns:
\itemize{
\item \code{.chain}. MCMC chain index.
\item \code{.term}. Parameter information.
\item \code{.plot}. List column of MCMC visualisations.
}
}

\seealso{
Other drawsvis methods: 
\code{\link{drawsvis_to_patchwork}()}
}
\concept{drawsvis methods}
