% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawsvis-to-patchwork.R
\name{drawsvis_to_patchwork}
\alias{drawsvis_to_patchwork}
\title{Drawsvis to Patchwork}
\usage{
drawsvis_to_patchwork(dv, ..., separate_chains = TRUE)
}
\arguments{
\item{dv}{Drawsvis-type tibble. See \strong{Drawsvis} section.}

\item{...}{extra arguments passed to \code{\link[patchwork:wrap_plots]{wrap_plots}}.}

\item{separate_chains}{logical. Input is split across \code{.chain} when \code{TRUE}.}
}
\value{
A \code{\link[patchwork:patchwork-package]{patchwork}} object or a list of
such objects when \code{separate_chains} is \code{TRUE}.
}
\description{
Convert to a \code{\link[patchwork:patchwork-package]{patchwork}} object.
}
\section{Drawsvis}{

Typically output from \code{\link[=drawsvis]{drawsvis()}}, a \code{drawsvis}-style
\link[tibble:tibble]{tibble} has one term per row instead of one draws
per row and is therefore assumed to have columns:
\itemize{
\item \code{.chain}. MCMC chain index.
\item \code{.term}. Parameter information.
\item \code{.plot}. List column of MCMC visualisations.
}
}

\seealso{
Other drawsvis methods: 
\code{\link{drawsvis_add_term_subtitle}()}
}
\concept{drawsvis methods}
