% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo-mcmcrutils}
\alias{loo-mcmcrutils}
\alias{log_pred_density}
\alias{waic.mcmcarray}
\alias{loo.mcmcarray}
\alias{loo_summary}
\alias{loo_compare_summary}
\alias{loo_compare_summary_print}
\title{LOO Cross Validation within \code{mcmcr}}
\usage{
log_pred_density(x, separate_chains = TRUE)

waic.mcmcarray(x, separate_chains = TRUE, ...)

loo.mcmcarray(x, separate_chains = TRUE, r_eff = NULL, ...)

loo_summary(x, separate_chains = TRUE, r_eff = NULL)

loo_compare_summary(loglik_list, separate_chains = TRUE, r_eff = NULL)

loo_compare_summary_print(x, digits = 2, type = c("gt", "latex"))
}
\arguments{
\item{x}{\code{\link[mcmcr:mcmcarray-object]{mcmcarray}} object. Log-likelihood.}

\item{separate_chains}{logical. Chain analysed independently when \code{TRUE}.}

\item{...}{These dots are for future extensions and must be empty.}

\item{r_eff}{vector. Either \code{"auto"} to uses \code{\link[loo:relative_eff]{loo::relative_eff()}} or will
be passed directly to \code{\link[loo:loo]{loo::loo()}}.}

\item{loglik_list}{named list of \code{\link[mcmcr:mcmcarray-object]{mcmcarray}}
objects representing different model's log-likelihood.}

\item{digits}{integer. Number of decimal places for all numeric quantities.}

\item{type}{choice. Return type, either a \code{gt} object or LaTeX code.}
}
\description{
Methods for \code{\link[loo:loo-package]{loo()}} package functionality.
}
\details{
\code{log_pred_density()} is the simulation-estimate for log pointwise
predictive density (LPD).
}
\section{Functions}{
\itemize{
\item \code{loo_summary()}: Creates a 1-row tibble of each LOO metric (LPD, WAIC, PSIS).

\item \code{loo_compare_summary()}: Creates a tibble with one row per model showing metric comparisons and
comparsion of standard errors, see \code{\link[loo:loo_compare]{loo_compare}}.

\item \code{loo_compare_summary_print()}: Output \code{\link[=loo_compare_summary]{loo_compare_summary()}} as a human-readable table.

}}
\references{
Vehtari et al. (2017), \url{https://doi.org/10.1007/s11222-016-9696-4}.
}
\seealso{
LPD (see references), PSIS (\code{\link[loo:loo]{loo::loo()}}),
WAIC (\code{\link[loo:waic]{loo::waic()}}).
}
